+++ ESFramework/Core/ESFrameworkCore.h
@@ -12,6 +12,7 @@
 #import "NSObject+ESAssociatedObjectHelper.h"
 #import "NSObject+ESAutoCoding.h"
 #import "NSRegularExpression+ESAdditions.h"
+#import "NSCharacterSet+ESAdditions.h"
 #import "NSString+ESAdditions.h"
 #import "NSString+ESGTMHTML.h"
 #import "NSString+ESHash.h"