+++ CHANGELOG.md
@@ -18,6 +18,7 @@
 - Removed `ESOSVersionIsAbove*()`, `ESStringFromSize()`, `NSStringWith()`, `UIImageFromCache()`, `UIImageFrom()` functions.
 - Removed `UIAlertView+ESBlock`, `UIActionSheet+ESBlock`.
 - Renamed `ESTouchDirectoryAtURL()` to `ESTouchDirectoryAtFileURL()`.
+- Added `+[NSCharacterSet URLEncodingAllowedCharacterSet]`
 
 ## 2.6.2